Series
The program runs in series.  It contains log2 transform, outlier removal using the cdf test, quantile normalization, and the KS test.  

Before running the file, be sure to change the input GEM file and its path on line 4 and the output .txt file and its path on line 99.  

Type qsub norm_series.pbs in the command line to run.