Parallel
This file runs in parallel.  It contains everything in the series program (norm_series.py) except for quantile normalization.  This program includes log2 transform, outlier removal using the CDF test, and the KS test.

Before running the file, be sure to change the input GEM file and its path on line 8 and the output .txt file and its path on line 64.  

The output of this program includes the gene IDs and the p and d values.